♻️Refactor: BookInfo 컴포넌트로 추출, 중복코드 제거, 폴더 정리
