+++ src/Components/layout/BookInfo.js
@@ -1,0 +1,70 @@
+import React from "react";
+import styled from "styled-components";
+import theme from "../../Styles/theme";
+
+import { useLocation } from "react-router-dom";
+import CircleIcon from "@material-ui/icons/RadioButtonUnchecked";
+
+/////////////////////////////// 책 상세정보를 보여주는 컴포넌트. ListItem에서 선택하여 state로 보낸 책 데이터를 받는다
+const BookInfo = () => {
+  const location = useLocation();
+
+  return (
+    <div className="contents_div">
+      <Div className="rowDirection">
+        <img src={location.state.thumbnail} alt={location.state.thumbnail} />
+        <DivBookInfo>
+          <H1>{location.state.title}</H1>
+          <div className="rowDirection">
+            <H3>{location.state.authors}</H3>
+            <CircleIcon
+              style={{
+                color: `${theme.colors.peacockGray}`,
+                fontSize: "8px",
+                marginRight: "2%",
+              }}
+            />
+            <p className="subText">
+              {location.state.publisher} ({location.state.publishDate})
+            </p>
+          </div>
+          <br /> <Hr /> <br />
+          <DivBookArticle>
+            {location.state.contents && (
+              <article>{location.state.contents + " ..."}</article>
+            )}
+          </DivBookArticle>
+        </DivBookInfo>
+      </Div>
+    </div>
+  );
+};
+
+const Div = styled.div`
+  margin: 20px;
+  width: 100%;
+`;
+const DivBookInfo = styled.div`
+  margin-left: 50px;
+  width: 100%;
+  height: auto;
+`;
+const DivBookArticle = styled.div`
+  height: 100px;
+  //overflow-y: scroll;
+`;
+
+const Hr = styled.hr`
+  background-color: ${theme.colors.lightGray};
+  border-width: 1px 0px 0px 0px;
+  height: 1px;
+`;
+const H1 = styled.h1`
+  color: ${theme.colors.peacock};
+`;
+const H3 = styled.h3`
+  color: ${theme.colors.peacockGray};
+  margin-right: 2%;
+`;
+
+export default BookInfo;
