+++ src/Pages/SearchBook/BookDetail.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import BookInfo from "../../Components/layout/BookInfo";
+
+const BookDetail = () => {
+  return (
+    <div className="contents_div">
+      <BookInfo />
+    </div>
+  );
+};
+
+export default BookDetail;
